+++ wallet_helper/__init__.py
@@ -23,7 +23,7 @@ from wallet_helper.ledger import Ledger, LedgerLike, make_key
 from wallet_helper.remote import RemoteLedger
 from wallet_helper.sqlite_ledger import SqliteLedger
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 
 __all__ = [
     "Ledger",